+++ ED_AP.py
@@ -8,6 +8,7 @@ import cv2
 from PIL import Image
 from pathlib import Path
 
+from EDAP_data import *
 from EDlogger import logger, logging
 import Image_Templates
 import Screen
@@ -18,6 +19,7 @@ from EDKeys import *
 from EDafk_combat import AFK_Combat
 from ModulesInfoParser import ModulesInfoParser
 from Overlay import *
+from StatusParser import StatusParser
 from Voice import *
 from Robigo import *
 
@@ -121,6 +123,7 @@ class EDAutopilot:
         self.waypoint = EDWayPoint(self.jn.ship_state()['odyssey'])
         self.robigo = Robigo(self)
         self.modules_info = ModulesInfoParser()
+        self.status = StatusParser()
 
         # rate as ship dependent.   Can be found on the outfitting page for the ship.  However, it looks like supercruise
         # has worse performance for these rates
@@ -462,23 +465,44 @@ class EDAutopilot:
         else:
             return True
 
-    def being_interdicted(self, scr_reg) -> bool:
-        interdict_image, (minVal, maxVal, minLoc, maxLoc), match = scr_reg.match_template_in_region('interdicted', 'interdicted')
-
+    def interdiction_check(self) -> bool:
+        """ Checks if we are being interdicted. This can occur in SC and maybe in system jump by Thargoids
+        (needs to be verified). Returns False if not interdicted, True after interdiction is detected and we
+        get away. Use return result to determine the next action (continue, or do something else).
         """
-        if self.cv_view:
-            #self.draw_match_rect(interdict_image, pt, (pt[0] + width, pt[1] + height), (0,255,0), 2)
-            cv2.putText(interdict_image, f'{maxVal:5.2f} >.40', (1, 20), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 1, cv2.LINE_AA)
-            cv2.imshow('interdict', interdict_image)
-            cv2.moveWindow('interdict', self.cv_view_x-460,self.cv_view_y+560)
-            cv2.waitKey(1)
-        """
-
-        # need > 50% in the match to say we do have a interdiction
-        if maxVal < 0.40:
+        # Return if we are not being interdicted.
+        if not self.status.get_flag(FlagsBeingInterdicted):
             return False
-        else:
-            return True
+
+        # Interdiction detected.
+        self.vce.say("Danger. Interdiction detected.")
+        self.ap_ckb('log', 'Interdiction detected.')
+
+        # Keep setting speed to zero to submit while in supercruise or system jump.
+        while self.status.get_flag(FlagsSupercruise) or self.status.get_flag2(Flags2FsdHyperdriveCharging):
+            self.keys.send('SetSpeedZero')  # Submit.
+            sleep(0.5)
+
+        # Set speed to 100%.
+        self.keys.send('SetSpeed100')
+
+        # Wait for cooldown to start.
+        self.status.wait_for_flag_on(FlagsFsdCooldown)
+
+        # Boost while waiting for cooldown to complete.
+        while not self.status.wait_for_flag_off(FlagsFsdCooldown, timeout=2):
+            self.keys.send('UseBoostJuice')
+
+        # Cooldown over, get us out of here.
+        self.keys.send('HyperSuperCombination', hold=0.001)
+
+        # Wait for supercruise, keep boosting.
+        while not self.status.wait_for_flag_on(FlagsSupercruise, timeout=2):
+            self.keys.send('UseBoostJuice')
+
+        # Update journal flag.
+        self.jn.ship_state()['interdicted'] = False  # reset flag
+        return True
 
     def get_nav_offset(self, scr_reg):
         """ Determine the x,y offset from center of the compass of the nav point. """
@@ -643,6 +667,12 @@ class EDAutopilot:
     #  
     def undock(self):
         # Assume we are in Star Port Services                              
+        # Go to interior panel which will update ModulesInfo.json if any modules have changed.
+        self.keys.send('UIFocus', state=1)
+        self.keys.send('UI_Right')
+        self.keys.send('UIFocus', state=0)
+        self.keys.send('UI_Back')
+
         # Now we are on initial menu, we go up to top (which is Refuel)
         self.keys.send('UI_Up', repeat=3)
 
@@ -772,6 +802,13 @@ class EDAutopilot:
         # if sun in front of us, then keep pitching up until it is below us
         while self.is_sun_dead_ahead(scr_reg):
             self.keys.send('PitchUpButton', state=1)
+
+            # check if we are being interdicted
+            interdicted = self.interdiction_check()
+            if interdicted:
+                # Continue journey after interdiction
+                self.keys.send('SetSpeedZero')
+
             # if we are pitching more than N seconds break, may be in high density area star area (close to core)
             if ((time.time()-starttime) > fail_safe_timeout):
                 logger.debug('sun avoid failsafe timeout')
@@ -1176,9 +1213,9 @@ class EDAutopilot:
             scr_reg.set_sun_threshold(self.config['SunBrightThreshold'])
                     
         # Lets avoid the sun, shall we
-        self.vce.say("Sun avoidance")
-        self.update_ap_status("Circumnavigating star")
-        self.ap_ckb('log', 'Circumnavigating star')
+        self.vce.say("Avoiding star")
+        self.update_ap_status("Avoiding star")
+        self.ap_ckb('log', 'Avoiding star')
         self.sun_avoid(scr_reg)
 
         if self.jn.ship_state()['fuel_percent'] < self.config['RefuelThreshold'] and is_star_scoopable and has_fuel_scoop:
@@ -1200,6 +1237,12 @@ class EDAutopilot:
             #if we don't scoop first 5 tons with 40 sec break, since not scooping or not fast enough or not at all, then abort
             startime = time.time()
             while not self.jn.ship_state()['is_scooping'] and not self.jn.ship_state()['fuel_percent'] == 100:
+                # check if we are being interdicted
+                interdicted = self.interdiction_check()
+                if interdicted:
+                    # Continue journey after interdiction
+                    self.keys.send('SetSpeedZero')
+
                 if ((time.time()-startime) > int(self.config['FuelScoopTimeOut'])):
                     self.vce.say("Refueling abort, insufficient scooping")
                     return False
@@ -1209,6 +1252,12 @@ class EDAutopilot:
             # We started fueling, so lets give it another timeout period to fuel up
             startime = time.time()
             while not self.jn.ship_state()['fuel_percent'] == 100:
+                # check if we are being interdicted
+                interdicted = self.interdiction_check()
+                if interdicted:
+                    # Continue journey after interdiction
+                    self.keys.send('SetSpeedZero')
+
                 if ((time.time()-startime) > int(self.config['FuelScoopTimeOut'])):
                     self.vce.say("Refueling abort, insufficient scooping")
                     return True
@@ -1433,7 +1482,6 @@ class EDAutopilot:
         # Loop forever keeping tight align to target, until we get SC Disengage popup
         while True:
             sleep(0.05)
-
             if self.jn.ship_state()['status'] == 'in_supercruise':
 
                 # Align and stay on target. If false is returned, we have lost the target behind us.
@@ -1448,21 +1496,12 @@ class EDAutopilot:
                 align_failed = True
                 break
 
-            # check if we are being interdicted, means we saw it on screen or we already got interdicted
-            if self.being_interdicted(scr_reg) == True or self.jn.ship_state()['interdicted'] == True:
-
-                self.keys.send('SetSpeedZero')  # submit
-                sleep(2)  # wait to come out of interdiction screen
-                self.keys.send('SetSpeed100')
-
-                while self.jn.ship_state()['status'] == 'in_space':
-                    self.keys.send('UseBoostJuice')
-                    self.keys.send('HyperSuperCombination', hold=0.001)
-                    sleep(5)
-
-                self.nav_align(scr_reg)  # realign with station
+            # check if we are being interdicted
+            interdicted = self.interdiction_check()
+            if interdicted:
+                # Continue journey after interdiction
                 self.keys.send('SetSpeed50')
-                self.jn.ship_state()['interdicted'] = False  # reset flag
+                self.nav_align(scr_reg)  # realign with station
 
             # check for SC Disengage
             if (self.sc_disengage(scr_reg) == True):