+++ Screen.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+import typing
 import logging
 
 import cv2
@@ -26,6 +28,7 @@ Author: sumzer0@yahoo.com
 #     img = ImageGrab.grab(bbox)
 
 elite_dangerous_window = "Elite - Dangerous (CLIENT)"
+
 
 class Screen:
     def __init__(self):
@@ -64,8 +67,8 @@ class Screen:
             mon_num = mon_num + 1
 
         # Add new screen resolutions here with tested scale factors
-        # this table will be default, overwriten when loading resolution.json file
-        self.scales = {  #scaleX, scaleY
+        # this table will be default, overwritten when loading resolution.json file
+        self.scales = {  # scaleX, scaleY
             '1024x768':   [0.39, 0.39],  # tested, but not has high match % 
             '1080x1080':  [0.5, 0.5],    # fix, not tested
             '1280x800':   [0.48, 0.48],  # tested
@@ -98,7 +101,7 @@ class Screen:
             self.scaleY = self.scales[scale_key][1]
         except:            
             # if we don't have a definition for the resolution then use calculation
-            self.scaleX = self.screen_width  / 3440.0
+            self.scaleX = self.screen_width / 3440.0
             self.scaleY = self.screen_height / 1440.0
             
         # if the calibration scale values are not -1, then use those regardless of above
@@ -111,7 +114,7 @@ class Screen:
         logger.debug('Scale X, Y: '+str(self.scaleX)+", "+str(self.scaleY))
 
     @staticmethod
-    def get_elite_window_rect() -> tuple[int, int, int, int] | None:
+    def get_elite_window_rect() -> typing.Tuple[int, int, int, int] | None:
         """ Gets the ED window rectangle.
         Returns (left, top, right, bottom) or None.
         """
@@ -144,9 +147,7 @@ class Screen:
 
     # reg defines a box as a percentage of screen width and height
     def get_screen_region(self, reg):      
-        image = array(self.mss.grab((int(reg[0]), int(reg[1]), 
-                                        int(reg[2]), int(reg[3]) )))
-        image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
+        image = self.get_screen(int(reg[0]), int(reg[1]), int(reg[2]), int(reg[3]))
         return image
     
     def get_screen(self, x_left, y_top, x_right, y_bot):    #  if absolute need to scale??